requirements.txt
pyswisseph>=2.10
yfinance>=0.2.28
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
plotly>=5.15
streamlit>=1.28
fpdf2>=2.7
schedule>=1.2
python-dateutil>=2.8
pytz>=2023.3